+++ CHANGES.rst
@@ -14,6 +14,8 @@ Changes
 * Updated test environment: Added PyPy, PyPy3 and Python 3.6
 * QRCode.get_png_size() is deprecated, use QRCode.symbol_size(). The latter
   returns a (width, height) tuple, not an integer.
+* Deprecated QRCode.png_as_base64_str(), use QRCode.png_data_uri() which returns
+  a valid URI
 
 
 1.2.1 - 2016-06-20